--- DotNet/AccountTests/AccountTestsConstants.cs
@@ -1,9 +1,0 @@
-﻿namespace AccountUnitTests
-{
-    public static class AccountTestsConstants
-    {
-        public const string email = "testAccount@gmail.com";
-        public const string accountId = "6B29FC40-CA47-1067-B31D-00DD010662DA";
-        public const string roleId = "6B29FC40-CA47-1067-D70B-99BB101448BC";
-    }
-}
