--- DotNet/AccountTests/TestAsyncEnumerable.cs
@@ -1,34 +1,0 @@
-﻿using System.Linq.Expressions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AccountUnitTests
-{
-    public class TestAsyncEnumerable<T> : EnumerableQuery<T>, IAsyncEnumerable<T>, IQueryable<T>
-    {
-        public TestAsyncEnumerable(IEnumerable<T> enumerable) : base (enumerable)
-        {
-
-        }
-
-        public TestAsyncEnumerable(Expression expression) : base(expression)
-        {
-            
-        }
-
-        public IAsyncEnumerator<T> GetEnumerator()
-        {
-            return new TestAsyncEnumerator<T>(this.AsEnumerable().GetEnumerator());
-        }
-
-        IQueryProvider IQueryable.Provider => new TestAsyncQueryProvider<T>(this);
-
-        public IAsyncEnumerator<T> GetAsyncEnumerator(CancellationToken cancellationToken)
-        {
-            return new TestAsyncEnumerator<T>(this.AsEnumerable().GetEnumerator());
-        }
-    } 
-}
