+++ DotNet/ServiceTests/UnitTests/Account/TestAsyncEnumerable.cs
@@ -1,0 +1,30 @@
+﻿using System.Linq.Expressions;
+using Expression = System.Linq.Expressions.Expression;
+
+namespace UnitTests.Account
+{
+    public class TestAsyncEnumerable<T> : EnumerableQuery<T>, IAsyncEnumerable<T>, IQueryable<T>
+    {
+        public TestAsyncEnumerable(IEnumerable<T> enumerable) : base (enumerable)
+        {
+
+        }
+
+        public TestAsyncEnumerable(Expression expression) : base(expression)
+        {
+            
+        }
+
+        public IAsyncEnumerator<T> GetEnumerator()
+        {
+            return new TestAsyncEnumerator<T>(this.AsEnumerable().GetEnumerator());
+        }
+
+        IQueryProvider IQueryable.Provider => new TestAsyncQueryProvider<T>(this);
+
+        public IAsyncEnumerator<T> GetAsyncEnumerator(CancellationToken cancellationToken)
+        {
+            return new TestAsyncEnumerator<T>(this.AsEnumerable().GetEnumerator());
+        }
+    } 
+}
